+++ Lucy/Controllers/RegComidaController.cs
@@ -62,7 +62,7 @@ namespace Lucy.Controllers
             ViewBag.lComidas = new SelectList(lComidas, "Id", "Valor");
 
 
-            int idUsu = Fachada.Functions.get_idUsu(Request.Cookies[FormsAuthentication.FormsCookieName]);
+            long idUsu = Fachada.Functions.get_idUsu(Request.Cookies[FormsAuthentication.FormsCookieName]);
 
             List<ModelCL.Alimento> lAlimentos = db.Alimento.Where(a => a.Usuario == null || a.Usuario.UsuarioId == idUsu).ToList();
             newRegComida.Alimentos = new List<ComidaAlimentoViewModel>();
@@ -175,7 +175,7 @@ namespace Lucy.Controllers
             ViewBag.lComidas = new SelectList(lComidas, "Id", "Valor");
 
 
-            int idUsu = Fachada.Functions.get_idUsu(Request.Cookies[FormsAuthentication.FormsCookieName]);
+            long idUsu = Fachada.Functions.get_idUsu(Request.Cookies[FormsAuthentication.FormsCookieName]);
 
             List<ModelCL.Alimento> lAlimentos = db.Alimento.Where(a => a.Usuario == null || a.Usuario.UsuarioId == idUsu).ToList();
             vmRegComida.Alimentos = new List<ComidaAlimentoViewModel>();
